+++ languagetool-core/src/main/java/org/languagetool/rules/neuralnetwork/IClassifier.java
@@ -2,6 +2,6 @@ package org.languagetool.rules.neuralnetwork;
 
 public interface IClassifier {
 
-    double[] getScores(String[] context);
+    float[] getScores(String[] context);
 
 }
