+++ languagetool-core/src/test/java/org/languagetool/rules/neuralnetwork/MatrixTest.java
@@ -11,24 +11,24 @@ public class MatrixTest {
     @Test
     public void matrixFromListTest() {
         String matrixString = "1 2\n3 4\n5 6";
-        Matrix expectedMatrix = new Matrix(new double[][]{{1,2},{3,4},{5,6}});
+        Matrix expectedMatrix = new Matrix(new float[][]{{1,2},{3,4},{5,6}});
         Matrix matrix = new Matrix(Arrays.asList(matrixString.split("\n")));
         assertEquals(expectedMatrix, matrix);
     }
 
     @Test
     public void matMulTest() {
-        final Matrix a = new Matrix(new double[][]{{1,2},{3,4},{5,6}});
-        final Matrix b = new Matrix(new double[][]{{1},{2}});
-        final Matrix c = new Matrix(new double[][]{{5},{11},{17}});
+        final Matrix a = new Matrix(new float[][]{{1,2},{3,4},{5,6}});
+        final Matrix b = new Matrix(new float[][]{{1},{2}});
+        final Matrix c = new Matrix(new float[][]{{5},{11},{17}});
         assertEquals(c, a.mul(b));
     }
 
     @Test
     public void matAddTest() {
-        Matrix a = new Matrix(new double[][]{{1,2},{3,4},{5,6}});
-        Matrix b = new Matrix(new double[][]{{1,2},{3,4},{5,7}});
-        Matrix c = new Matrix(new double[][]{{2,4},{6,8},{10,13}});
+        Matrix a = new Matrix(new float[][]{{1,2},{3,4},{5,6}});
+        Matrix b = new Matrix(new float[][]{{1,2},{3,4},{5,7}});
+        Matrix c = new Matrix(new float[][]{{2,4},{6,8},{10,13}});
         assertEquals(c, a.add(b));
     }
